Finished refactoring api solution
