+++ back/MyStore.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.1.32421.90
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.Web", "MyStore.Web\MyStore.Web.csproj", "{4AC5F8EC-5FCB-4AC8-9C02-AB59EEBCFC1F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.UnitTests", "MyStore.UnitTests\MyStore.UnitTests.csproj", "{EA4F5C85-037C-454B-8E3C-CCF1C7463323}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.Core.Repository", "MyStore.Core.Repository\MyStore.Core.Repository.csproj", "{13932FE9-5375-468F-9643-B0F016295BB1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.Core.Data.Entity.Dto", "MyStore.Core.Data\MyStore.Core.Data.Entity\MyStore.Core.Data.Entity.Dto\MyStore.Core.Data.Entity.Dto.csproj", "{47FEAE96-586C-43B0-9667-4D6854871190}"
@@ -25,9 +23,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.Core.Domain.Service
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{56856E44-1B36-410D-B2C3-34BA1641DED0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyStore.Core.Data.Entity.Relation.Test.Unit", "Test\MyStore.Core.Data.Entity.Relation.Test.Unit\MyStore.Core.Data.Entity.Relation.Test.Unit.csproj", "{A8B19CEA-6700-4609-A6C2-63446EEDEB18}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.Core.Data.Entity.Relation.Test.Unit", "Test\MyStore.Core.Data.Entity.Relation.Test.Unit\MyStore.Core.Data.Entity.Relation.Test.Unit.csproj", "{A8B19CEA-6700-4609-A6C2-63446EEDEB18}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyStore.Core.Repository.Test.Unit", "Test\MyStore.Core.Repository.Test.Unit\MyStore.Core.Repository.Test.Unit.csproj", "{021E41C7-52A3-4A8F-AEC2-2ABB12AC9D3F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyStore.Core.Repository.Test.Unit", "Test\MyStore.Core.Repository.Test.Unit\MyStore.Core.Repository.Test.Unit.csproj", "{021E41C7-52A3-4A8F-AEC2-2ABB12AC9D3F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,10 +37,6 @@ Global
 		{4AC5F8EC-5FCB-4AC8-9C02-AB59EEBCFC1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4AC5F8EC-5FCB-4AC8-9C02-AB59EEBCFC1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4AC5F8EC-5FCB-4AC8-9C02-AB59EEBCFC1F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA4F5C85-037C-454B-8E3C-CCF1C7463323}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA4F5C85-037C-454B-8E3C-CCF1C7463323}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA4F5C85-037C-454B-8E3C-CCF1C7463323}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA4F5C85-037C-454B-8E3C-CCF1C7463323}.Release|Any CPU.Build.0 = Release|Any CPU
 		{13932FE9-5375-468F-9643-B0F016295BB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{13932FE9-5375-468F-9643-B0F016295BB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{13932FE9-5375-468F-9643-B0F016295BB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
