+++ back/Test/MyStore.Core.Data.Entity.Relation.Test.Unit/ProductTests.cs
@@ -1,58 +1,11 @@
-using MyStore.Core.Data.Entity.Model;
-using MyStore.Core.Repository;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using MyStore.Core.Data.Entity.Relation;
 using Xunit;
-using Xunit.Abstractions;
-using Xunit.Microsoft.DependencyInjection.Attributes;
 
-namespace MyStore.UnitTests
+namespace MyStore.Core.Repository.Test.Unit
 {
-    [TestCaseOrderer("Xunit.Microsoft.DependencyInjection.TestsOrder.TestPriorityOrderer", "Xunit.Microsoft.DependencyInjection")]
-    [CollectionDefinition("Products")]
-    [Collection("Products")]
-    public class ProductTests : BaseTests
+    public class ProductTests
     {
-        private readonly IStoreRepository<Product> _productRepository;
-
-        public ProductTests(ITestOutputHelper testOutputHelper, UnitTestsFixture fixture) : base(testOutputHelper, fixture)
-            => _productRepository = fixture.GetService<IStoreRepository<Product>>(testOutputHelper);
-
-        [Fact, TestOrder(1)]
-        public void WhenZeroProduct_DoesNotReturnNull()
-        {
-            List<Product> actualProducts = _productRepository.GetAll();
-
-            Assert.NotNull(actualProducts);
-        }
-
-        [Fact, TestOrder(2)]
-        public void WhenZeroProduct_ReturnsEmpty()
-        {
-            List<Product> actualProducts = _productRepository.GetAll();
-
-            Assert.Empty(actualProducts);
-        }
-
-        [Fact, TestOrder(3)]
-        public async Task WhenAtLeastOneProduct_CanGetAllProducts()
-        {
-            var expectedProducts = new List<Product>
-            {
-                new Product { Name = "product1", Description = "product1Description", Price = 100 },
-                new Product { Name = "product2", Description = "product2Description", Price = 200 },
-                new Product { Name = "product3", Description = "product3Description", Price = 300 }
-            };
-            await _context.AddRangeAsync(expectedProducts);
-            await _context.SaveChangesAsync();
-
-            List<Product> actualProducts = _productRepository.GetAll();
-
-            Assert.True(expectedProducts.SequenceEqual(actualProducts));
-        }
-
-        [Fact, TestOrder(4)]
+        [Fact]
         public void TwoSameProducts_EqualEachOther()
         {
             var stubProduct1 = GetProduct();
@@ -63,7 +16,7 @@ namespace MyStore.UnitTests
             Assert.True(areEqual);
         }
 
-        [Fact, TestOrder(5)]
+        [Fact]
         public void TwoProducts_WhenDifferentIds_DoNotEqualEachOther()
         {
             var stubProduct1 = GetProduct();
@@ -74,7 +27,7 @@ namespace MyStore.UnitTests
             Assert.False(areEqual);
         }
 
-        [Fact, TestOrder(6)]
+        [Fact]
         public void TwoProducts_WhenDifferentNames_DoNotEqualEachOther()
         {
             var stubProduct1 = GetProduct();
@@ -86,7 +39,7 @@ namespace MyStore.UnitTests
             Assert.False(areEqual);
         }
 
-        [Fact, TestOrder(7)]
+        [Fact]
         public void TwoProducts_WhenDifferentDescriptions_DoNotEqualEachOther()
         {
             var stubProduct1 = GetProduct();
@@ -98,7 +51,7 @@ namespace MyStore.UnitTests
             Assert.False(areEqual);
         }
 
-        [Fact, TestOrder(8)]
+        [Fact]
         public void TwoProducts_WhenDifferentPrices_DoNotEqualEachOther()
         {
             var stubProduct1 = GetProduct();
@@ -110,7 +63,7 @@ namespace MyStore.UnitTests
             Assert.False(areEqual);
         }
 
-        [Fact, TestOrder(9)]
+        [Fact]
         public void WhenSecondProductIsNull_DoNotEqualEachOther()
         {
             var stubProduct1 = GetProduct();
