+++ back/Test/MyStore.Core.Data.Test.Unit.Common/TestsOrderer.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace MyStore.Core.Data.Test.Unit.Common
+{
+    public class TestsOrderer : ITestCollectionOrderer
+    {
+        public IEnumerable<ITestCollection> OrderTestCollections(IEnumerable<ITestCollection> testCollections)
+            => testCollections.OrderBy(collection => collection.DisplayName);
+    }
+}
